Add Staff form
This form accomplishes stories 2 (display a staff register) and 3 (display a staff members timetable), as they're closely related. The idea being, the user will select the date to view a register for (idea: should it default to today on form load?) and from the register view (key-value behind the scenes of staffId and Name, but only show name) select the staff member they wish to see the timetable for in the list and click timetable. In the timetable listbox (display time as key and booked or free for value) on the same date they previously selected.