Development of staff form
So far I have replaced the monthcalendar with a datetimepicker so that it was possible for the application to understand days without complicated coding. I have also added some if statements so that the staff form doesn't connect to the database if the date selected is a saturday or sunday.