Tweaks to date select and database connection start
I've made a few tweaks to the date selection so if the user selects a saturday or sunday the view timetable button is disabled because there just won't be any data to show. I've also tried to make a start on the database connection but I still need to understand connections more to make greater progress.