Ongoing patient profile development
So far I have added functionality to the view profile button on the patient view and the back button on the profile view. The prescription dgv shows all available medications however we want to show only the medication prescribed to the patient. Having a few difficulties with data transfer from search dvg to profile labels and also producing patient tests.